+++ lesson3/pom.xml
@@ -34,7 +34,11 @@
             <version>4.0.0</version>
             <scope>provided</scope>
         </dependency>
-
+                <dependency>
+                    <groupId>com.fasterxml.jackson.core</groupId>
+                    <artifactId>jackson-databind</artifactId>
+                    <version>2.11.1</version>
+                </dependency>
 
         <dependency>
             <groupId>javax.annotation</groupId>
