+++ lesson3/src/main/webapp/WEB-INF/views/my/home.html
@@ -3,14 +3,16 @@
 
 <head>
     <meta charset="ISO-8859-1">
-    <title>My json</title>
+    <title>My home</title>
 </head>
 
 <body>
 
-<p>json</p>
-<h1 th:text="'text: ' + ${receivedText}" />
+<p>home</p>
 
+
+<a href="/json/get">get json</a>
+<a href="/xml/get">get xml</a>
 </body>
 
 </html>
